refactor: move syncCloudSaves from GOGService to GOGManager
GOGService.syncCloudSaves was a static companion method that needed
to acquire a service instance to reach GOGManager. Moving it to
GOGManager eliminates the indirection and makes GOGCloudSavesManager
an injected dependency rather than an ad-hoc instantiation per sync.

GOGService.syncCloudSaves is now a thin delegate.
